feat: add `fieldsBefore` and `fieldsAfter` ApiResource extenders
